Extract facets, etc
https://coveord.atlassian.net/browse/KIT-1622
